refactor: 🔨 Scroll.ts 파일 이동
